requirements.txt
# for crypto
python-gnupg

# for compression
pip2 install pylzma
